wrong solve

+++ B_Print.c
@@ -1,12 +1,14 @@
 #include <stdio.h>
-void print(long long int x){
-    for(int i=1;i<=x;i++){
-        printf("%d ",i);
+
+void print(long long int x) {
+    for (long long int i = 1; i <= x; i++) {
+        printf("%lld ", i);
     }
 }
+
 int main() {
     long long int x;
-    scanf("%lld",&x);
+    scanf("%lld", &x);
     print(x);
     return 0;
 }
